Add additional warning if the SDK version specified does not match what was resolved
Add unit tests
