+++ src/Build.UnitTests/Microsoft.Build.Engine.UnitTests.csproj
@@ -105,7 +105,7 @@
     <Compile Include="BackEnd\MockLoggingService.cs" />
     <Compile Include="BackEnd\MockTaskBuilder.cs" />
     <Compile Include="BackEnd\MSBuild_Tests.cs" />
-    <Compile Include="BackEnd\SdkResolution_Tests.cs" />
+    <Compile Include="BackEnd\SdkResolverService_Tests.cs" />
     <Compile Include="BackEnd\SdkResolverLoader_Tests.cs" />
     <Compile Include="BackEnd\TargetBuilder_Tests.cs" />
     <Compile Include="BackEnd\EventRedirectorToSink_Tests.cs " />
